requirements.txt
IMDbPY==2022.7.9
python-dotenv==1.0.0
Requests==2.31.0
tmdbv3api==1.9.0